# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Onereach API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "abiodun.adeyinka@telvida.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/apponereach/public",
            "description": "Onereach API Server"
        }
    ],
    "paths": {
...
        "/api/app/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Current Authenticated User",
                "description": "Get Current Authenticated User",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "User details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/api/app/media/pilot-number": {
            "get": {
                "tags": [
                    "Media Pilot Number"
                ],
                "summary": "Get Pilot Numbers",
                "description": "Returns Pilot Number Attached To Current Tenant",
                "operationId": "getPilotNumber",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
        "/api/app/pilot/search": {
            "get": {
                "tags": [
                    "Pilot Number"
                ],
                "summary": "Search pilot number",
                "description": "Returns list of available pilot numbers",
                "operationId": "getPilotNumber",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "phone number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/pilot/reserve": {
            "get": {
                "tags": [
                    "Pilot Number"
                ],
                "summary": "Add pilot number to cart ",
                "description": "Returns pilot number for few hours until cart expires",
                "operationId": "reservePilotNumberById",
                "parameters": [
                    {
                        "name": "number_id",
                        "in": "path",
                        "description": "pilot number id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/pilot/remove": {
            "get": {
                "tags": [
                    "Pilot Number"
                ],
                "summary": "Remove pilot number from cart",
                "description": "Return remaining list of numbers in the cart",
                "operationId": "removePilotNumberById",
                "parameters": [
                    {
                        "name": "number_id",
                        "in": "path",
                        "description": "Pilot Number id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/pilot/reserved": {
            "get": {
                "tags": [
                    "Pilot Number"
                ],
                "summary": "Get Number is the cart",
                "description": "Returns list of pilot numbers in the cart.",
                "operationId": "getReservedNumbers",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/trackers": {
            "get": {
                "tags": [
                    "User Tracker"
                ],
                "summary": "Get Tracker",
                "description": "Return Tracker.",
                "operationId": "getTracker",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/trackers/check-in": {
            "post": {
                "tags": [
                    "User Tracker"
                ],
                "summary": "Log user longitude and latitude",
                "description": "Check-in user longitude and lat",
                "operationId": "checkInLocation",
                "parameters": [
                    {
                        "name": "long",
                        "in": "path",
                        "description": "Longitude",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "path",
                        "description": "Latitude",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Action have been logged"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/trackers/user": {
            "get": {
                "tags": [
                    "User Tracker"
                ],
                "summary": "Get User Tracker",
                "description": "Return Tracker.",
                "operationId": "getUserTracker",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
...
}